+++ packages/eyes-selenium/CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Unreleased
 
 
+## 4.53.0 - 2021/3/29
+
+- updated to @applitools/eyes-sdk-core@12.17.0 (from 12.16.2)
+- updated to @applitools/visual-grid-client@15.7.0 (from 15.6.2)
+
 ## 4.52.8 - 2021/3/16
 
 - updated to @applitools/visual-grid-client@15.6.0 (from 15.5.22)